app/src/main/java/com/example/tingxie/domain/use_case/GetCharacters.kt
package com.example.tingxie.domain.use_case

import com.example.tingxie.domain.model.Character
import com.example.tingxie.domain.repository.CharacterRepository
import kotlinx.coroutines.flow.Flow
import javax.inject.Inject

class GetCharacters (
    private val characterRepository: CharacterRepository
) {
    fun getCharacters(): Flow<List<Character>> {
        // Can do ordering here
        return characterRepository.getCharacters()
    }

    suspend fun getCharacter(id: Int): Character? {
        return characterRepository.getCharacter(id)
    }

    fun getNRandomCharacters(number: Int): Flow<List<Character>> {
        return characterRepository.getNRandomCharacters(number)
    }
}